Remove locality stuff and more cleanup
